Fixed warnings found with -Wshorten-64-to-32 flag. 	Change on 2014/06/18 by tball <tball@google.com> ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=69469620

+++ jre_emul/Classes/IOSArray.m
@@ -42,7 +42,7 @@
 
   __unsafe_unretained IOSClass *componentTypes[dimensionCount];
   componentTypes[dimensionCount - 1] = nil;
-  for (int i = dimensionCount - 2; i >= 0; i--) {
+  for (NSInteger i = (NSInteger) dimensionCount - 2; i >= 0; i--) {
     __unsafe_unretained IOSClass *last = componentTypes[i + 1];
     if (last) {
       componentTypes[i] = [IOSClass arrayClassWithComponentType:last];
